refactor(qr,rewards): extract qr/[code] and rewards pages onto features/ (TASK-21854)
state machines and tier math move into hooks, four duplicated qr return
branches collapse into parameterized views; route pages become thin
wrappers per the features/home pattern. the code query fallback moves to
nuqs (new-file ratchet). zero behavior change otherwise.
